--add new level commit

+++ src/the_bank/deposit_boxes/Valuable.java
@@ -1,8 +1,7 @@
 package the_bank.deposit_boxes;
 
 
-import java.math.BigDecimal;
-
 public interface Valuable {
-    BigDecimal getValue();
+    double getValue();
+
 }
